Day 3

+++ app/Http/Controllers/noteController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Note;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class noteController extends Controller
 {
@@ -21,6 +22,7 @@ class noteController extends Controller
 
     public function show(Note $note)
     {
+        Gate::authorize('view', $note);
         $note->load('category');
         $note->created_human = $note->created_at->diffForHumans();
         return inertia('note/show', compact('note'));
@@ -28,34 +30,33 @@ class noteController extends Controller
 
     public function create()
     {
-        $user = request()->user();
-        $categories = $user->categories()->get();
-
+        $categories = Auth::user()->categories()->get();
+        
         return inertia('note/create', compact('categories'));
     }
-
+    
     public function post()
     {
-        $user = request()->user();
-
         $validated = request()->validate([
             'title' => ['required', 'string', 'min:5', 'max:128'],
             'content' => ['required', 'string', 'min:20', 'max:5000'],
             'category_id' => ['required', 'exists:categories,id']
         ]);
 
-        $user->notes()->create($validated);
+        Auth::user()->notes()->create($validated);
 
         return redirect()->route('home');
     }
 
     public function edit(Note $note)
     {
+        Gate::authorize('view', $note);
         $categories = Auth::user()->categories()->get();
         return inertia('note/edit', compact('note', 'categories'));
     }
 
     public function update(Note $note) {
+        Gate::authorize('update', $note);
         $validated = request()->validate([
             'title' => ['required', 'string', 'min:5', 'max:128'],
             'content' => ['required', 'string', 'min:20', 'max:5000'],
@@ -69,6 +70,7 @@ class noteController extends Controller
 
     public function delete(Note $note)
     {
+        Gate::authorize('delete', $note);
         $note->delete();
 
         return redirect()->route('home');
